Extract expression dependency observer
Move dependency collection and change detection out of useExpressionDataSources so the hook is responsible for wiring React state and snapshots, while ExpressionObserver owns dependency tracking, subscriptions, and rerender scheduling.

Build the expression runtime through a dedicated factory and pass current snapshots through a ref. This removes the previous inline observer implementation and the artificial runtimeRevision dependency trick, making the data flow easier to follow without changing runtime behavior.
